+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Corrected documentation badge label in READE.MD - fixes [issue #95](https://github.com/IAG-NZ/jenkins/issues/95).
+- Added basic docker support for creation of integration tests.
 
 ## 1.0.2.240